+++ support-frontend/assets/pages/paper-subscription-landing/components/hero/heroPicture.jsx
@@ -1,0 +1,31 @@
+// @flow
+
+import React from 'react';
+import GridPicture from 'components/gridPicture/gridPicture';
+
+const HeroPicture = () => (
+  <GridPicture
+    sources={[
+      {
+        gridId: 'paperLandingHeroMobile',
+        srcSizes: [500, 922],
+        imgType: 'png',
+        sizes: '100vw',
+        media: '(max-width: 739px)',
+      },
+      {
+        gridId: 'paperLandingHero',
+        srcSizes: [1000, 2000],
+        imgType: 'png',
+        sizes: '(min-width: 1000px) 2000px, 1000px',
+        media: '(min-width: 740px)',
+      },
+    ]}
+    fallback="paperLandingHero"
+    fallbackSize={1000}
+    altText=""
+    fallbackImgType="png"
+  />
+);
+
+export { HeroPicture };
